requirements.txt
ruamel.yaml
pillow
pandas
opencv-python
imgaug
imagecorruptions